pyproject.toml
[tool.poetry]
name = "cometd-poc-py"
version = "0.1.0"
description = ""
authors = ["Toby <toby@thedevlins.biz>"]

[tool.poetry.dependencies]
python = "^3.8"
pyjwt = "2.0.1" # breaking api changes in 2.1.0
aiosfstream = "^0.5.0"
cryptography = "^3.4.7"

[tool.poetry.dev-dependencies]
black = "^21.4b2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
